Support reading/writing data to/from a static PHP array
If $wgTrustedXffFile ends with '.php', assume it is a PHP file that returns a
static associative array. This is to provide a nominal performance improvement
on HHVM, which caches the bytecode of PHP files, making relying on them more
efficient than hitting the disk for the cdb file each time.

Bug: T141120
Change-Id: Ia1976b91a897f7fd80f3ddf5a142d441ac0e88aa