Do not fail when trusted-xff.cdb does not exist
Sometime one might not have generated the cdb file. Throw an exception
and catch it in the hook. A debug log is added for the sysadmin.
The IP remains untrusted.

Also add a test to ensure the extension sets a sane default for
$wgTrustedXffFile.

Bug: T198120
Change-Id: I485f3f5931e5c2b1a7829d8733c68174c6ae26de
(cherry picked from commit d0b070c200940b1140831daae87bd073ab6b0944)