Remove some NXDOMAIN DNS entries
Not a valid host or IP address on line 397: cache01.nyc.untd.com.
Not a valid host or IP address on line 398: cache02.nyc.untd.com.
Not a valid host or IP address on line 399: cache03.nyc.untd.com.
Not a valid host or IP address on line 400: cache04.nyc.untd.com.

Change-Id: I15e1bb5bd54241ac30985e0e7d7101113db794cb